+++ packages/ormed/lib/src/builder/model_helper.dart
@@ -59,8 +59,7 @@ String modelHelperClass(String className) {
   buffer.writeln('  static ModelFactoryBuilder<$className> factory({');
   buffer.writeln('    GeneratorProvider? generatorProvider,');
   buffer.writeln('  }) =>');
-  buffer.writeln('      ModelFactoryBuilder<$className>(');
-  buffer.writeln('        definition: definition,');
+  buffer.writeln('      ModelFactoryRegistry.factoryFor<$className>(');
   buffer.writeln('        generatorProvider: generatorProvider,');
   buffer.writeln('      );');
   buffer.writeln();